Move window-manager, rename storage route module for consistency
