+++ client/src/entry/analysis/router.js
@@ -55,7 +55,7 @@ import Register from "entry/analysis/modules/Register";
 import WorkflowEditorModule from "entry/analysis/modules/WorkflowEditor";
 import AdminRoutes from "entry/analysis/routes/admin-routes";
 import LibraryRoutes from "entry/analysis/routes/library-routes";
-import StorageDashboardRoutes from "entry/analysis/routes/storageDashboardRoutes";
+import StorageRoutes from "entry/analysis/routes/storage-routes";
 import { getAppRoot } from "onload/loadConfig";
 import Vue from "vue";
 import VueRouter from "vue-router";
@@ -220,7 +220,7 @@ export function getRouter(Galaxy) {
                 children: [
                     ...AdminRoutes,
                     ...LibraryRoutes,
-                    ...StorageDashboardRoutes,
+                    ...StorageRoutes,
                     {
                         path: "",
                         alias: "root",
