+++ client/src/entry/analysis/routes/storage-routes.ts
