+++ client/src/entry/analysis/window-manager.js
